+++ src/index.ts
@@ -253,9 +253,9 @@ export * from './identity/ownership/TokenOwnershipValidator';
 
 // Identity/Storage
 export * from './identity/storage/AdapterFactory';
+export * from './identity/storage/ClientIdAdapterFactory';
 export * from './identity/storage/ExpiringAdapterFactory';
 export * from './identity/storage/PassthroughAdapterFactory';
-export * from './identity/storage/WebIdAdapterFactory';
 
 // Identity
 export * from './identity/AccountInitializer';